converted number to the proper type needed by client
